tests: factor out common code
We will reuse this for image-boot tests.
